fix(@formatjs/intl-datetimeformat): fix mutability issue

+++ packages/intl-datetimeformat/src/skeleton.ts
@@ -78,7 +78,7 @@ export function parseDateTimeSkeleton(skeleton: string): Formats {
 
         // Weekday
         case 'E':
-          result.weekday = len === 4 ? 'short' : len === 5 ? 'narrow' : 'short';
+          result.weekday = len === 4 ? 'long' : len === 5 ? 'narrow' : 'short';
           return '{weekday}';
 
         case 'e':
